refactor: move access_token endpoints to dedicated api file
